requirements.txt
streamlit
reportlab
pillow
pandas
gspread
oauth2client
python-dotenv
